backend/requirements.txt
AgentLogin==3.0.0
amqp==5.0.2
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.3.0
captcha==0.3
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.10
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
Deprecated==1.2.13
Django==2.2
django-cors-headers==3.11.0
django-crontab==0.7.1
django-haystack==3.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
elasticsearch==2.4.1
eventlet
greenlet==0.4.17
httplib2==0.20.2
idna==2.10
itsdangerous==1.1.0
Jinja2
kombu==5.0.2
MarkupSafe==1.1.1
mutagen==1.45.1
mysqlclient==2.0.2
packaging==21.3
pika==1.1.0
Pillow
prompt-toolkit==3.0.8
py3Fdfs==2.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyJWT
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.6
python-alipay-sdk==2.1.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
